requirements.txt
# Core NLP
transformers==4.40.0
datasets==2.19.0
tokenizers==0.19.0
torch==2.3.0
accelerate==0.30.0

# Evaluation
seqeval==1.2.2
scikit-learn==1.4.2
bert-score==0.3.13

# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
httpx==0.27.0

# LLM (rewriter)
anthropic==0.26.0
openai==1.30.0

# UI
streamlit==1.35.0

# Utils
python-dotenv==1.0.1
huggingface-hub==0.23.0
